Implement retriving product details functionality
