+++ src/am.kon.projects.marktguru-test/Controllers/ProductsController.cs
@@ -11,8 +11,8 @@ using Microsoft.AspNetCore.Mvc;
 namespace am.kon.projects.marktguru_test.Controllers;
 
 [Authorize]
-[ResponseCache(Duration = 20, Location = ResponseCacheLocation.Any)]
-public class ProductsController : Controller
+//[ResponseCache(Duration = 20, Location = ResponseCacheLocation.Any)]
+public partial class ProductsController : Controller
 {
     private readonly ILogger<ProductsController> _logger;
     private readonly ProductManagementService _productManagementService;
@@ -59,51 +59,5 @@ public class ProductsController : Controller
         }
 
         return View(pageModel);
-    }
-
-    [HttpGet]
-    public async Task<IActionResult> CreateProduct()
-    {
-        return View();
-    }
-
-    [HttpPost]
-    [ValidateAntiForgeryToken]
-    public async Task<IActionResult> CreateProduct(ProductCreateModel productCreateModel)
-    {
-        if (!ModelState.IsValid)
-        {
-            return View(productCreateModel);
-        }
-
-        Product product = productCreateModel.ToProduct();
-
-        try
-        {
-            ProductActionResult<Product> result = await _productManagementService.Create(product);
-            
-            switch (result.ActionResult)
-            {
-                case ProductActionResultTypes.Ok:
-                    return RedirectToAction("Index");
-                
-                case ProductActionResultTypes.Error:
-                case ProductActionResultTypes.Info:
-                    ModelState.AddModelError("error-info", result.Message);
-                    break;
-                
-                default:
-                    ModelState.AddModelError("unhadled-case", "Unhandled case during product item creation.");
-                    break;
-            }
-
-        }
-        catch (Exception ex)
-        {
-            ModelState.AddModelError("unhadled-exception", "Unhandled exception during product item creation.");
-            _logger.LogDebug(ex, "Unhandled exception.");
-        }
-
-        return View();
     }
 }
