+++ src/am.kon.projects.marktguru-test/appsettings.Development.json
@@ -1,4 +1,7 @@
 {
+  "ConnectionStrings": {
+    "DefaultConnection": "Server=hs01.kon.am;Database=am.kon.projects.marktguru-test;User ID=am.kon.projects.marktguru-test-user;Password=-n6Al1P-faMq=3gbzo8v4_pwmKpT.mu;Trust Server Certificate =Yes;"
+  },
   "Logging": {
     "LogLevel": {
       "Default": "Information",
